requirements.txt
flask
pytest
Flask
mysql-connector-python